Handle object being destroyed before endResize debounce
A component that utilizes this mixin might be destroyed shortly after resizing, but before `_endResize` has finished debouncing. This results in the `_endResize` function being run when the object is already destroyed, causing an `Uncaught Error: Assertion Failed: calling set on destroyed object` when it tries to set `resizing` to false.